streamlined_requirements.txt
# Core dependencies only - optimized for free hosting
Flask==2.3.0
Flask-CORS==4.0.0
google-generativeai==0.3.0
numpy==1.21.0
scipy==1.9.0
requests==2.31.0
cryptography==41.0.0
python-dotenv==1.0.0

# Minimal additional dependencies
pgmpy==0.1.20
